requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.12.0
flask-migrate==4.0.5
httpx==0.24.1
Werkzeug==3.0.1 